Fix issues #1 and #2
Use dateutil.parser to avoid "time data does not match format" error. This
was due to datetime.strptime not being able to parse PDT or any TZ field
correctly if $LC_TIME wasn't set. parser simply ignores the field, which we
don't need explicitly in this situation.

Also fix the dict comprehension in __init__.py which isn't suppored in Python 2.6.